Refactoring: Move PermissionResourceImpl into UserResourceImpl and delete PermissionResourceImpl
