+++ manager/api/micro/src/main/java/io/apiman/manager/api/micro/ManagerApiMicroServiceApplication.java
@@ -51,7 +51,6 @@ public class ManagerApiMicroServiceApplication extends ApiManagerApplication {
         classes.add(RoleResourceImpl.class);
         classes.add(UserResourceImpl.class);
         classes.add(CurrentUserResourceImpl.class);
-        classes.add(PermissionsResourceImpl.class);
         classes.add(OrganizationResourceImpl.class);
         classes.add(PolicyDefinitionResourceImpl.class);
         classes.add(GatewayResourceImpl.class);
