+++ manager/api/war/wildfly8/src/main/webapp/WEB-INF/web.xml
@@ -118,15 +118,6 @@
     </web-resource-collection>
     <web-resource-collection>
       <web-resource-name>apiman</web-resource-name>
-      <url-pattern>/permissions/*</url-pattern>
-      <http-method>GET</http-method>
-      <http-method>POST</http-method>
-      <http-method>PUT</http-method>
-      <http-method>DELETE</http-method>
-      <http-method>HEAD</http-method>
-    </web-resource-collection>
-    <web-resource-collection>
-      <web-resource-name>apiman</web-resource-name>
       <url-pattern>/plugins/*</url-pattern>
       <http-method>GET</http-method>
       <http-method>POST</http-method>
