+++ manager/test/api/src/test/resources/test-plan-data/idm/memberships/005_get-permissions.resttest
@@ -1,4 +1,4 @@
-GET /permissions/user1 admin/admin
+GET /users/user1/permissions admin/admin
 ----
 200
 Content-Type: application/json
