+++ manager/test/api/src/test/resources/test-plan-data/idm/memberships/007_get-permissions-user2-not-allowed.resttest
@@ -1,0 +1,9 @@
+GET /users/user1/permissions user2/user2
+----
+403
+Content-Type: application/json
+
+{"type":"NotAuthorizedException","errorCode":-1,"message":"Access denied.","moreInfoUrl":null}
+
+
+
